(#99) Add servlets to handle log + ZIP download
- refactor file + streaming handling into shared utilities
- abstract differences between servlet / web script via facade interface
- adjust Aikau service to handle additional URL type
- update / correct some copyright notes
- add explicit Surf dependency as transitive dependency was missing
- move dependency management information into appropriate section
