+++ share/pom.xml
@@ -22,21 +22,39 @@
         <dependency.surf.version>6.2</dependency.surf.version>
     </properties>
 
+    <dependencyManagement>
+        <dependencies>
+            <!-- Alfresco artifacts include bad transitive version -->
+            <dependency>
+                <groupId>org.alfresco.surf</groupId>
+                <artifactId>spring-surf</artifactId>
+                <version>${dependency.surf.version}</version>
+                <scope>provided</scope>
+            </dependency>
+
+            <dependency>
+                <groupId>${alfresco.groupId}</groupId>
+                <artifactId>aikau</artifactId>
+                <version>1.0.101</version>
+                <!-- we require user to pick the Aikau version they want to use - we just state we build and test against this version -->
+                <scope>provided</scope>
+                <exclusions>
+                    <!-- Alfresco is incapable of properly including transitive dependency versions -->
+                    <exclusion>
+                        <groupId>org.alfresco.surf</groupId>
+                        <artifactId>spring-surf-api</artifactId>
+                    </exclusion>
+                </exclusions>
+            </dependency>
+
+        </dependencies>
+    </dependencyManagement>
+
     <dependencies>
 
         <dependency>
             <groupId>${alfresco.groupId}</groupId>
             <artifactId>aikau</artifactId>
-            <version>1.0.101</version>
-            <!-- we require user to pick the Aikau version they want to use - we just state we build and test against this version -->
-            <scope>provided</scope>
-            <exclusions>
-                <!-- Alfresco is incapable of properly including transitive dependency versions -->
-                <exclusion>
-                    <groupId>org.alfresco.surf</groupId>
-                    <artifactId>spring-surf-api</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
 
         <dependency>
@@ -60,6 +78,11 @@
         <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-compress</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.alfresco.surf</groupId>
+            <artifactId>spring-surf</artifactId>
         </dependency>
 
     </dependencies>
