+++ share/src/main/amp/web/ootbee-support-tools/service/LogFileService.js
@@ -1,19 +1,26 @@
 /**
- * Copyright (C) 2016 Axel Faust Copyright (C) 2016 Order of the Bee
- * 
+ * Copyright (C) 2016, 2017 Axel Faust / Markus Joos
+ * Copyright (C) 2016, 2017 Order of the Bee
+ *
  * This file is part of Community Support Tools
- * 
- * Community Support Tools is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
- * 
- * Community Support Tools is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
- * 
- * You should have received a copy of the GNU Lesser General Public License along with Community Support Tools. If not, see
+ *
+ * Community Support Tools is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * Community Support Tools is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Community Support Tools. If not, see
  * <http://www.gnu.org/licenses/>.
  */
 /*
- * Linked to Alfresco Copyright (C) 2005-2016 Alfresco Software Limited.
+ * Linked to Alfresco
+ * Copyright (C) 2005-2017 Alfresco Software Limited.
  */
 
 /* global define: false */
@@ -47,6 +54,10 @@ define([ 'dojo/_base/declare', 'alfresco/services/BaseService', 'alfresco/core/C
             else if (payload.urlType === 'SHARE')
             {
                 baseUrl = Constants.URL_SERVICECONTEXT;
+            }
+            else if (payload.urlType === 'CONTEXT')
+            {
+                baseUrl = Constants.URL_CONTEXT;
             }
 
             if (baseUrl !== undefined && lang.isString(payload.baseUrl))
@@ -97,6 +108,10 @@ define([ 'dojo/_base/declare', 'alfresco/services/BaseService', 'alfresco/core/C
             {
                 baseUrl = Constants.URL_SERVICECONTEXT;
             }
+            else if (payload.urlType === 'CONTEXT')
+            {
+                baseUrl = Constants.URL_CONTEXT;
+            }
 
             path = this._toPath(payload);
 
@@ -133,6 +148,10 @@ define([ 'dojo/_base/declare', 'alfresco/services/BaseService', 'alfresco/core/C
             else if (payload.urlType === 'SHARE')
             {
                 baseUrl = Constants.URL_SERVICECONTEXT;
+            }
+            else if (payload.urlType === 'CONTEXT')
+            {
+                baseUrl = Constants.URL_CONTEXT;
             }
 
             path = this._toPath(payload);
