+++ share/src/main/resources/META-INF/web-fragment.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<web-fragment xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
+    xsi:schemaLocation="http://java.sun.com/xml/ns/javaee/web-fragment_3_0.xsd" id="WebAppFragment_ID" version="3.0">
+
+    <name>${moduleId}-web-fragment</name>
+
+    <ordering>
+        <after>
+            <others />
+        </after>
+    </ordering>
+
+    <servlet>
+        <servlet-name>OOTBee Support Tools - Log File Download</servlet-name>
+        <servlet-class>org.orderofthebee.addons.support.tools.share.LogFileGetServlet</servlet-class>
+    </servlet>
+    
+    <servlet>
+        <servlet-name>OOTBee Support Tools - Log ZIP File Download</servlet-name>
+        <servlet-class>org.orderofthebee.addons.support.tools.share.LogFilesZIPPostServlet</servlet-class>
+    </servlet>
+
+    <servlet-mapping>
+        <servlet-name>OOTBee Support Tools - Log File Download</servlet-name>
+        <url-pattern>/ootbee-support-tools/log4j-log-file/*</url-pattern>
+    </servlet-mapping>
+    
+    <servlet-mapping>
+        <servlet-name>OOTBee Support Tools - Log ZIP File Download</servlet-name>
+        <url-pattern>/ootbee-support-tools/log4j-log-files.zip</url-pattern>
+    </servlet-mapping>
+
+</web-fragment>
